pyproject.toml
[project]
name = "selltools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "asgiref==3.9.1",
    "certifi==2025.8.3",
    "cffi==2.0.0",
    "charset-normalizer==3.4.3",
    "crispy-bootstrap4==2025.6",
    "cryptography==45.0.7",
    "django==5.2",
    "django-allauth==65.11.2",
    "django-cors-headers>=4.9.0",
    "django-crispy-forms==2.4",
    "django-easy-faq>=1.9",
    "django-model-utils==5.0.0",
    "django-phonenumber-field==8.1.0",
    "django-render-block==0.11",
    "djangorestframework>=3.16.1",
    "idna==3.10",
    "mysqlclient==2.2.7",
    "oauthlib==3.3.1",
    "phonenumbers==9.0.13",
    "pillow>=11.3.0",
    "pip>=25.3",
    "pycparser==2.23",
    "pyjwt==2.10.1",
    "python-dotenv>=1.1.1",
    "requests==2.32.5",
    "sqlparse==0.5.3",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "websockets>=15.0.1",
]
